improve delete routes

+++ tvali/tvali/api/routes/v1/crud/log.py
@@ -57,15 +57,17 @@ class LogCRUDFactory(BaseModel):
             query_params = query.model_dump(
                 exclude_none=True, exclude=["order_by", "limit", "offset"]
             )
-            filter_params = {
-                getattr(self.db_class, k): v for k, v in query_params.items()
-            }
+            filter_params = [
+                getattr(self.db_class, k) == v for k, v in query_params.items()
+            ]
 
             with session.get_db() as db:
                 try:
+                    print(self.db_class, filter_params, query)
                     objs = (
-                        db.query(self.db_class)
-                        .filter(**filter_params)
+                        db
+                        .query(self.db_class)
+                        .filter(*filter_params)
                         .limit(query.limit)
                         .offset(query.offset)
                         .all()
@@ -91,6 +93,12 @@ class LogCRUDFactory(BaseModel):
             with session.get_db() as db:
                 try:
                     item = db.query(self.db_class).filter(self.db_class.id == id_).first()
+
+                    if not item:
+                        raise HTTPException(
+                            status_code=404,
+                            detail=f"Could not find {self.tier.capitalize()}{self.log_type.capitalize()} object with id {id_}"
+                        )
                     db.delete(item)
 
                     db.commit()
